feat: Make $i18next basically reactive
This allows using $i18next.initialized and other properties
in the template and have Vue re-render it when some i18next
event (see the 'rerenderOn' option)  happens.
Technically it works like the $t() re-rendering, using a 'ref'.

We use a JS Proxy to intercept all property reads on $i18next.
This works only in mordern browsers, but that's in line with Vue 3's
own browser requirements.