feat: Make $i18next basically reactive
This allows using $i18next.isInitialized and other properties
in the template and have Vue re-render it when some i18next
event (see the 'rerenderOn' option) happens.
Technically it works like the $t() re-rendering, using an 'observable'.

We use a JS Proxy to intercept all property reads on $i18next.
This works only in mordern browsers. If there is no proxy support
(like in IE), we fall back to the original non-reactive i18next object.