feat: Translation component with slot interpolation
For Vue 2 we need to create VNode-s for the text nodes. This API is not
publicly available. We access it via the Vue._v function. This is internal, but
as Vue 2 will not get new minor releases, this is unlikely to go away.

By using a functional component we can render multiple root VNodes which
is otherwise not possible in Vue 2.

implements #8 for Vue 2