feat: Re-introduce support for i18nOptions in components
We now support i18nOptions with lng/keyPrefix/namespaces when
the plugin is initialized with the new
legacyI18nOptionsSupport: true
option.
This affects $t in the component, but not useTranslation()
like in version 2.

We dropped support for i18nOptions "messages" without a replacement.

This is really only a legacy option to allow converting components
incrementally over to useTranslation() with the respective options
to get a specific t() function.

We'll completely remove the i18nOptions support in i18next-vue v4.

The legacy option also restores support for having a keyPrefix in
i18nOptions but ignoring that for a "someNs:full.key" translation key.
This behaviour only applies to $t and will be removed in v4.